chore: separate graph, node and edge mapping
