+++ cmd/graph_test.go
@@ -1,0 +1,20 @@
+package cmd
+
+import (
+	"testing"
+
+	"github.com/elliotchance/orderedmap/v2"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestEmptyGraphHasNoNodes(t *testing.T) {
+	g := mkGraph(orderedmap.NewOrderedMap[string, []string]())
+
+	assert.Equal(t, 0, len(g.nodes))
+}
+
+func TestEmptyGraphHasNoEdges(t *testing.T) {
+	g := mkGraph(orderedmap.NewOrderedMap[string, []string]())
+
+	assert.Equal(t, 0, len(g.edges))
+}
